Fixed some bugs
FIxed a bug where if you throw a spider mine re grab it with the alt fire, and than throw it with the drop mine key when you had an other weapon selected, it wouldn't use ammo.
Also Changed some death messages, which may or may not gonna be desplayed, but did it anyways.